requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.12
fastapi-admin==1.0.4
greenlet==3.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
limits==5.0.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy-extensions==1.0.0
ngrok==1.4.0
openai==1.75.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
piccolo==1.24.2
piccolo_admin==1.9.1
piccolo_api==1.5.2
pillow==11.2.1
platformdirs==4.3.7
priority==2.0.0
psycopg2-binary==2.9.10
py==1.11.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
repoze.lru==0.7
requests==2.32.3
retry==0.9.2
retrying==1.3.4
Routes==2.5.1
secure==1.0.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.40
starlette==0.46.2
sudo==1.0.0
svgwrite==1.4.3
targ==0.4.0
tenacity==9.1.2
tortoise-orm==0.25.0
tqdm==4.67.1
Tree==0.2.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
vercel==0.2.1
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.1.2
